+++ lib/arachni/audit_store.rb
@@ -21,6 +21,9 @@ class AuditStore
     # @return    [Hash]  Plugin results.
     attr_accessor :plugins
 
+    # @return    [Array<Issue>]  Logged issues.
+    attr_accessor :issues
+
     # @return    [Time]    The date and time when the scan started.
     attr_accessor :start_datetime
 
@@ -53,28 +56,6 @@ class AuditStore
     # @return   [Hash]
     def options=( options )
         @options = prepare_options( options )
-    end
-
-    # @param    [Array<Issue>]  issues  Logged issues.
-    # @return   [Array<Issue>]
-    #   Logged issues sorted and grouped into variations.
-    def issues=( issues )
-        @issues = {}
-        Issue.sort( prepare_variations( issues ) ).each do |issue|
-            @issues[issue.digest] = issue
-        end
-        self.issues
-    end
-
-    # @return    [Array<Issue>]  Logged issues.
-    def issues
-        @issues.values
-    end
-
-    # @param    [Issue#digest]  digest
-    # @return    [Issue]
-    def issue_by_digest( digest )
-        @issues[digest]
     end
 
     # Loads and returns an AuditStore object from file
@@ -144,22 +125,6 @@ class AuditStore
     # @return    [Hash]
     def prepare_options( options )
         options.to_hash.symbolize_keys( false )
-    end
-
-    # @param    [Array<Issue>]    issues
-    # @return    [Array<Issue>]
-    #   New array of Issues with populated {Issue#variations}.
-    #
-    # @see Issue#variations
-    def prepare_variations( issues )
-        new_issues = {}
-        issues.each do |issue|
-            id = issue.hash
-            new_issues[id] ||= issue.with_variations
-            new_issues[id].variations << issue.as_variation
-        end
-
-        new_issues.values
     end
 
     # @param    [String, Float, Integer]    seconds