+++ lib/arachni/rpc/server/framework.rb
@@ -227,7 +227,7 @@ class Framework < ::Arachni::Framework
     #   First variations of all discovered issues with generic info filled in
     #   from the parent.
     def issues
-        auditstore.issues.map { |issue| issue.variations.first.to_solo issue }
+        state.issues.map { |issue| issue.variations.first.to_solo issue }
     end
 
     # @return   [Array<Hash>]   {#issues} as an array of Hashes.