+++ lib/arachni/state.rb
@@ -1,0 +1,32 @@
+=begin
+    Copyright 2010-2014 Tasos Laskos <tasos.laskos@gmail.com>
+    All rights reserved.
+=end
+
+require_relative 'state/issues'
+
+module Arachni
+
+# @author Tasos "Zapotek" Laskos <tasos.laskos@gmail.com>
+class State
+
+    # {State} error namespace.
+    #
+    # All {State} errors inherit from and live under it.
+    #
+    # @author Tasos "Zapotek" Laskos <tasos.laskos@gmail.com>
+    class Error < Arachni::Error
+    end
+
+class <<self
+
+    attr_reader :issues
+
+    def reset
+        @issues = Issues.new
+    end
+
+end
+reset
+end
+end