+++ spec/arachni/check/auditor_spec.rb
@@ -336,7 +336,7 @@ describe Arachni::Check::Auditor do
     describe '#register_results' do
         it 'registers issues with the framework' do
             @auditor.register_results( [ Factory[:issue] ] )
-            @framework.checks.results.first.should == Factory[:issue]
+            @framework.state.issues.flatten.first.should == Factory[:issue]
         end
     end
 
@@ -350,14 +350,13 @@ describe Arachni::Check::Auditor do
         it 'logs an issue' do
             @auditor.log_issue( issue_data )
 
-            logged_issue = @framework.checks.results.first
+            logged_issue = @framework.state.issues.flatten.first
 
             logged_issue.to_h.tap do |h|
                 h[:page][:dom][:transitions].first.delete :time
             end.should eq issue.to_h.merge( referring_page: {
                 body: @auditor.page.body,
                 dom:  @auditor.page.dom.to_h.tap do |h|
-                    #h[:transitions].first.delete :time
                     h.delete :skip_states
                 end
             }).tap { |h| h[:page][:dom][:transitions].first.delete :time }
@@ -366,7 +365,7 @@ describe Arachni::Check::Auditor do
         it 'assigns a #referring_page' do
             @auditor.log_issue( issue_data )
 
-            logged_issue = @framework.checks.results.first
+            logged_issue = @framework.state.issues.flatten.first
             logged_issue.referring_page.should == @auditor.page
         end
     end
@@ -375,7 +374,7 @@ describe Arachni::Check::Auditor do
         it 'preserves the given remarks' do
             @auditor.log( issue_data )
 
-            logged_issue = @framework.checks.results.first
+            logged_issue = @framework.state.issues.flatten.first
             logged_issue.remarks.first.should be_any
         end
 
@@ -384,13 +383,13 @@ describe Arachni::Check::Auditor do
 
             it 'includes response data' do
                 @auditor.log( issue_data )
-                @framework.checks.results.first.response.should ==
+                @framework.state.issues.flatten.first.response.should ==
                     issue_data[:page].response
             end
 
             it 'includes request data' do
                 @auditor.log( issue_data )
-                @framework.checks.results.first.request.should ==
+                @framework.state.issues.flatten.first.request.should ==
                     issue_data[:page].request
             end
         end
@@ -400,12 +399,10 @@ describe Arachni::Check::Auditor do
                 issue_data.delete(:page)
                 @auditor.log( issue_data )
 
-                @framework.checks.results.first.page.body.should ==
-                    @auditor.page.body
-                @framework.checks.results.first.response.should ==
-                    @auditor.page.response
-                @framework.checks.results.first.request.should ==
-                    @auditor.page.request
+                issue = @framework.state.issues.flatten.first
+                issue.page.body.should == @auditor.page.body
+                issue.response.should == @auditor.page.response
+                issue.request.should == @auditor.page.request
             end
         end
     end
@@ -423,7 +420,7 @@ describe Arachni::Check::Auditor do
                 @auditor.load_page_from( @url + '/link' )
                 @auditor.audit( @seed )
                 @framework.http.run
-                @framework.checks.results.size.should == 1
+                @framework.state.issues.size.should == 1
             end
         end
 
@@ -432,8 +429,8 @@ describe Arachni::Check::Auditor do
                 @auditor.load_page_from( @url + '/link' )
                 @auditor.audit( { unix: @seed }, substring: @seed )
                 @framework.http.run
-                @framework.checks.results.size.should == 1
-                issue = @framework.checks.results.first
+                @framework.state.issues.size.should == 1
+                issue = @framework.state.issues.flatten.first
                 issue.platform_name.should == :unix
                 issue.platform_type.should == :os
             end
@@ -451,8 +448,8 @@ describe Arachni::Check::Auditor do
                             elements: [ Arachni::Element::Link ]
                          )
                         @framework.http.run
-                        @framework.checks.results.size.should == 1
-                        issue = @framework.checks.results.first
+                        @framework.state.issues.size.should == 1
+                        issue = @framework.state.issues.flatten.first
                         issue.vector.class.should == Arachni::Element::Link
                         issue.vector.affected_input_name.should == 'link_input'
                     end
@@ -464,8 +461,8 @@ describe Arachni::Check::Auditor do
                             elements: [ Arachni::Element::Form ]
                          )
                         @framework.http.run
-                        @framework.checks.results.size.should == 1
-                        issue = @framework.checks.results.first
+                        @framework.state.issues.size.should == 1
+                        issue = @framework.state.issues.flatten.first
                         issue.vector.class.should == Arachni::Element::Form
                         issue.vector.affected_input_name.should == 'form_input'
                     end
@@ -477,8 +474,8 @@ describe Arachni::Check::Auditor do
                             elements: [ Arachni::Element::Cookie ]
                          )
                         @framework.http.run
-                        @framework.checks.results.size.should == 1
-                        issue = @framework.checks.results.first
+                        @framework.state.issues.size.should == 1
+                        issue = @framework.state.issues.flatten.first
                         issue.vector.class.should == Arachni::Element::Cookie
                         issue.vector.affected_input_name.should == 'cookie_input'
                     end
@@ -489,8 +486,8 @@ describe Arachni::Check::Auditor do
                                         elements: [ Arachni::Element::Cookie ]
                         )
                         @framework.http.run
-                        @framework.checks.results.size.should == 1
-                        issue = @framework.checks.results.first
+                        @framework.state.issues.size.should == 1
+                        issue = @framework.state.issues.flatten.first
                         issue.vector.class.should == Arachni::Element::Cookie
                         issue.vector.affected_input_name.should == 'vulnerable'
                     end
@@ -503,8 +500,8 @@ describe Arachni::Check::Auditor do
                             elements: [ Arachni::Element::Header ]
                          )
                         @framework.http.run
-                        @framework.checks.results.size.should == 1
-                        issue = @framework.checks.results.first
+                        @framework.state.issues.size.should == 1
+                        issue = @framework.state.issues.flatten.first
                         issue.vector.class.should == Arachni::Element::Header
                         issue.vector.affected_input_name.should == 'Referer'
                     end
@@ -516,7 +513,7 @@ describe Arachni::Check::Auditor do
                             format: [ Arachni::Check::Auditor::Format::STRAIGHT ]
                          )
                         @framework.http.run
-                        @framework.checks.results.size.should == 4
+                        @framework.state.issues.size.should == 4
                     end
                 end
             end
@@ -541,7 +538,7 @@ describe Arachni::Check::Auditor do
                             @framework.http.run
                         end
 
-                        @framework.checks.results.find do |i|
+                        @framework.state.issues.flatten.find do |i|
                             i.vector.affected_input_name == 'you_made_it'
                         end.should be_true
                     end
@@ -566,7 +563,7 @@ describe Arachni::Check::Auditor do
                             @framework.http.run
                         end
 
-                        issue = issues.first
+                        issue = issues.flatten.first
                         issue.should be_true
                         issue.vector.class.should == Arachni::Element::Form
                         issue.vector.affected_input_name.should == 'you_made_it'