+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 - `Framework`
     - `#audit_page` -- Updated to perform DOM/JS/AJAX analysis on the page and
         feed DOM page snapshots and new paths back to the `Framework`.
+- Added `State` -- Stores and provides access to the system's state.
+    - `Issues` -- Stores logged `Issue` objects.
 - Removed the `Spider`.
     - The Framework has grown to encompass a process providing the same
         functionality as a result of `Browser` analysis.